+++ src/test/kotlin/com/ibrahimdans/i18n/plugin/ide/inspections/CodeHighlightingTestBase.kt
@@ -15,6 +15,8 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
 
 fun CodeInsightTestFixture.customHighlightingCheck(fileName: String, code: String, translationName: String, translation: String) = this.runWithConfig(Config(defaultNs = "translation")) {
+    val parent = translationName.substringBeforeLast("/", "")
+    if (parent.isNotEmpty()) this.tempDirFixture.findOrCreateDir(parent)
     this.addFileToProject(translationName, translation)
     this.configureByText(fileName, code)
     this.checkHighlighting(true, true, true, true)